# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise ROI Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Archive/delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/regions:
    get:
      operationId: listProjectSalesRegions
      tags: [regions]
      summary: List sales regions for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sales regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSalesRegion"
    post:
      operationId: addProjectSalesRegion
      tags: [regions]
      summary: Add a sales region to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSalesRegionInput"
      responses:
        "201":
          description: Created sales region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSalesRegion"
  /projects/{id}/regions/{regionId}:
    patch:
      operationId: updateProjectSalesRegion
      tags: [regions]
      summary: Update fx multiplier for a project sales region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSalesRegionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSalesRegion"
    delete:
      operationId: removeProjectSalesRegion
      tags: [regions]
      summary: Remove a sales region from a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /projects/{id}/audit-log:
    get:
      operationId: getProjectAuditLog
      tags: [audit]
      summary: Get audit log for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, region, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        region:
          type: string
        productCategory:
          type: ["string", "null"]
        investmentSize:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, review, approved, archived]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        salesRegions:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSalesRegion"
    ProjectSalesRegion:
      type: object
      required: [id, projectId, region, fxMultiplier]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        region:
          type: string
        fxMultiplier:
          type: number
    ProjectSalesRegionInput:
      type: object
      required: [region]
      properties:
        region:
          type: string
        fxMultiplier:
          type: number
    ProjectSalesRegionUpdate:
      type: object
      properties:
        fxMultiplier:
          type: number
    ProjectInput:
      type: object
      required: [name, region]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        region:
          type: string
        productCategory:
          type: string
        investmentSize:
          type: number
        createdBy:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        region:
          type: string
        productCategory:
          type: string
        investmentSize:
          type: number
        status:
          type: string
          enum: [draft, review, approved, archived]
...
    AuditLogEntry:
      type: object
      required: [id, tableName, recordId, fieldName, oldValue, newValue, changedBy, changedAt]
      properties:
        id:
          type: integer
        tableName:
          type: string
        recordId:
          type: integer
        fieldName:
          type: string
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        changedBy:
          type: string
        changedAt:
          type: string
          format: date-time
...